Fix confusion of sensors with SensorSocketAgent
When SensorSocketAgent is used, calling writeMsg immediately triggers msgSent which calls executeCommand again.
In the end the nested executeCommand call manages to enqueue its request to mProcessingFIFO before the request of the original call is enqueued.

Finally, when the response for the first sent command is received it is matched with the second command and vice-versa.
This resulted in sensor labels getting mixed up.